Switch to long

+++ src/Main.java
@@ -1,9 +1,9 @@
 public class Main {
 
     public static void main(String[] args) {
-        int[] sales = new int[]{7, 8, 777, 888};
+       long[] sales = new long[]{7, 8, 777, 888};
         SalesManager salesManager = new SalesManager(sales);
-        int max = salesManager.max();
+        long max = salesManager.max();
         System.out.println(max);
 
     }
